requirements.txt
ratelimit
cffi==1.9.1
contextlib2==0.5.4
cryptography==1.7.1
defusedxml==0.4.1
Django==1.10
django-allauth==0.30.0
django-braces==1.10.0
django-cors-headers==2.0.0
django-debug-panel==0.8.3
django-debug-toolbar==1.6
django-filter==0.15.2
django-graphiql==0.4.4
django-model-utils==2.6.1
django-oauth-toolkit==0.11.0
django-ratelimit==1.0.1
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.3
djangorestframework-oauth==1.1.0
Fabric==1.13.1
Fabric3==1.13.1.post1
graphene==1.1.3
graphene-django==1.2.1
graphql-core==1.0.1
graphql-django-view==1.3
graphql-relay==0.4.5
idna==2.2
iso8601==0.1.11
oauthlib==1.1.2
paramiko==2.1.1
pip-save==0.1.4
promise==0.4.2
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
python-openid==2.2.5
python3-openid==3.0.9
raven==5.32.0
requests==2.13.0
requests-oauthlib==0.7.0
singledispatch==3.4.0.3
six==1.10.0
social-auth-app-django==1.0.0
social-auth-core==1.0.0
sqlparse==0.2.2
stripe==1.46.0
wheel==0.26.0
jsondiff==1.1.0
pytest==3.0.7
factory-boy==2.8.1
pendulum==1.2.0
simplejson==3.10.0
